% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_binaries}
\alias{install_binaries}
\title{Install multiple binaries}
\usage{
install_binaries(filenames, lib = .libPaths()[[1L]],
  lock = getOption("install.lock", TRUE), num_workers = 1)
}
\arguments{
\item{filenames}{filenames of tarballs to install}

\item{lib}{library to install packages into}

\item{lock}{If any value but `FALSE`, use per-package locking when
installing. It defaults to using `getOption("install.lock")` for
compatibility with `utils::install.packages()`.}

\item{num_workers}{Number of parallel workers to use}
}
\description{
Install multiple binaries
}
