% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_binary.R
\name{install_binary}
\alias{install_binary}
\title{Install a R binary package}
\usage{
install_binary(filename, lib = .libPaths()[[1L]], metadata = NULL)
}
\arguments{
\item{filename}{filename of built binary package to install}

\item{lib}{library to install packages into}

\item{metadata}{Named character vector of metadata entries to be added
to the \code{DESCRIPTION} after installation.}
}
\description{
Install a R binary package
}
