% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_package}
\alias{install_package}
\title{Install a local R package}
\usage{
install_package(filename, lib = .libPaths()[[1L]], metadata = NULL,
  vignettes = TRUE, ...)
}
\arguments{
\item{filename}{filename of package to install. Can be a source
directory, source tarball or binary package.}

\item{lib}{library to install packages into}

\item{metadata}{Named character vector of metadata entries to be added
to the \code{DESCRIPTION} after installation.}

\item{vignettes}{whether to (re)build the vignettes of the packages.
It is ignored for binary packages.}

\item{...}{Additional arguments passed to \code{\link[pkgbuild:build]{pkgbuild::build()}}.}
}
\description{
Install a local R package
}
\keyword{internal}
