% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_source.R
\name{install_source}
\alias{install_source}
\title{Install a source package}
\usage{
install_source(path, lib = .libPaths()[[1L]],
  lock = getOption("install.lock", TRUE), quiet = TRUE, metadata = NULL,
  ...)
}
\arguments{
\item{path}{Path to a package, or within a package.}

\item{lib}{library to install packages into}

\item{lock}{If any value but `FALSE`, use per-package locking when
installing. It defaults to using `getOption("install.lock")` for
compatibility with `utils::install.packages()`.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}

\item{metadata}{Named character vector of metadata entries to be added
to the \code{DESCRIPTION} after installation.}

\item{...}{Additional arguments passed to [pkgbuild::build].}
}
\description{
Install a source package
}
