% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_binary.R
\name{verify_binary}
\alias{verify_binary}
\title{Verify a R binary archive is valid}
\usage{
verify_binary(filename)
}
\arguments{
\item{filename}{filename of built binary package to install}
}
\value{
A [desc::desc()] object of the DESCRIPTION file from the binary
  archive.
}
\description{
Verify a R binary archive is valid
}
