% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{install_binary}
\alias{install_binary}
\title{Install a R binary package}
\usage{
install_binary(filename, lib = .libPaths()[[1L]],
  lock = getOption("install.lock", TRUE))
}
\arguments{
\item{filename}{filename of built binary package to install}

\item{lib}{library to install packages into}

\item{lock}{If any value but `FALSE`, use per-package locking when
installing. It defaults to using `getOption("install.lock")` for
compatibility with `utils::install.packages()`.}
}
\description{
Install a R binary package
}
