% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_package}
\alias{install_package}
\title{Install a local R package}
\usage{
install_package(filename, lib = .libPaths()[[1L]],
  lock = getOption("install.lock", TRUE), ...)
}
\arguments{
\item{filename}{filename of package to install. Can be a source
directory, source tarball or binary package.}

\item{lib}{library to install packages into}

\item{lock}{If any value but `FALSE`, use per-package locking when
installing. It defaults to using `getOption("install.lock")` for
compatibility with `utils::install.packages()`.}

\item{...}{Additional arguments passed to [pkgbuild::build].}
}
\description{
Install a local R package
}
