% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_packages}
\alias{install_packages}
\title{Install multiple local packages}
\usage{
install_packages(filenames, lib = .libPaths()[[1L]],
  lock = getOption("install.lock", TRUE), num_workers = 1)
}
\arguments{
\item{filenames}{filenames of packages to install. Can be source
directories, source tarballs or binary packages.}

\item{lib}{library to install packages into}

\item{lock}{If any value but `FALSE`, use per-package locking when
installing. It defaults to using `getOption("install.lock")` for
compatibility with `utils::install.packages()`.}

\item{num_workers}{Number of parallel workers to use}
}
\description{
Install multiple local packages
}
