% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_source.R
\name{install_source}
\alias{install_source}
\title{Install a source package}
\usage{
install_source(path, lib = .libPaths()[[1L]], quiet = TRUE,
  metadata = NULL, vignettes = TRUE, ..., compile_attributes = NULL)
}
\arguments{
\item{path}{Path to a package, or within a package.}

\item{lib}{library to install packages into}

\item{quiet}{if \code{TRUE} suppresses output from this function.}

\item{metadata}{Named character vector of metadata entries to be added
to the \code{DESCRIPTION} after installation.}

\item{vignettes}{whether to (re)build the vignettes of the packages}

\item{...}{Additional arguments passed to \code{\link[pkgbuild:build]{pkgbuild::build()}}.}

\item{compile_attributes}{Passed on to \code{\link[pkgbuild:build]{pkgbuild::build()}}, default \code{FALSE}
if installing from tarball, \code{TRUE} otherwise.}
}
\description{
Supports installation from directories or tarballs.
}
